DPC-236: Reorganize converters
Split the common converters into Hibernate and Jackson packages, just to make things a bit clearer.
